Restructured project and extracted client base test class
