+++ Ravendarque.ClimbingCapacity/Ravendarque.ClimbingCapacity.Web.UnitTests/Ravendarque.ClimbingCapacity.Web.UnitTests.csproj
@@ -20,4 +20,9 @@
     <ProjectReference Include="..\Ravendarque.ClimbingCapacity.Web\Ravendarque.ClimbingCapacity.Web.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Clients\" />
+    <Folder Include="Parsers\" />
+  </ItemGroup>
+
 </Project>
