+++ Ravendarque.ClimbingCapacity/Ravendarque.ClimbingCapacity.Web/Parsers/ICapacityDataParser.cs
@@ -1,6 +1,6 @@
 ﻿using Ravendarque.ClimbingCapacity.Web.Models;
 
-namespace Ravendarque.ClimbingCapacity.Web.Services;
+namespace Ravendarque.ClimbingCapacity.Web.Parsers;
 
 public interface ICapacityDataParser<out T> where T : ICapacity
 {
