+++ Ravendarque.ClimbingCapacity/Ravendarque.ClimbingCapacity.Web/Ravendarque.ClimbingCapacity.Web.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="6.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Parsers\" />
+  </ItemGroup>
+
 </Project>
